Working on adding the Beta-era generator
- A few instances of uint8_t are now using BlockType for consistency
- Added all the necessary code for the Beta-era generator
- Terrain does load and in generated, though its not correct
- Updated all the parts of the generator to make use of the structs defined in numeric_structs

+++ src/bpp_shared/world/chunk_serializer.h
@@ -31,7 +31,7 @@ namespace ChunkSerializer {
                     bool highNibble = idx & 1;
                     Int3 pos{ x, y, z };
 
-                    blockData[idx] = chunk.getBlock(pos);
+                    blockData[idx] = uint8_t(chunk.getBlock(pos));
 
                     auto packNibble = [](uint8_t& byte, uint8_t val, bool high) {
                         if (high) byte = uint8_t((byte & (uint8_t(0x0F))) | ((val & uint8_t(0x0F)) << 4));
